+++ Tests/Functional/Localization/TranslateTraitTest.php
@@ -6,7 +6,10 @@ namespace OliverKlee\Seminars\Tests\Functional\Localization;
 
 use Nimut\TestingFramework\TestCase\FunctionalTestCase;
 use OliverKlee\Seminars\Localization\TranslateTrait;
+use TYPO3\CMS\Core\Information\Typo3Version;
 use TYPO3\CMS\Core\Localization\LanguageService;
+use TYPO3\CMS\Core\Localization\LanguageServiceFactory;
+use TYPO3\CMS\Core\Utility\GeneralUtility;
 
 /**
  * @covers \OliverKlee\Seminars\Localization\TranslateTrait
@@ -30,7 +33,11 @@ final class TranslateTraitTest extends FunctionalTestCase
 
     public function setUpLanguageService(): void
     {
-        $languageService = LanguageService::create('default');
+        if ((new Typo3Version())->getMajorVersion() >= 11) {
+            $languageService = GeneralUtility::makeInstance(LanguageServiceFactory::class)->create('default');
+        } else {
+            $languageService = LanguageService::create('default');
+        }
         $languageService->includeLLFile('EXT:seminars/Resources/Private/Language/locallang.xlf');
 
         $GLOBALS['LANG'] = $languageService;